+++ execution_engine/src/core/execution/executor.rs
@@ -48,13 +48,14 @@ macro_rules! on_fail_charge {
                 return ExecutionResult::Failure {
                     error: exec_err.into(),
                     execution_effect: Default::default(),
+                    execution_journal: Default::default(),
                     transfers: $transfers,
                     cost: $cost,
                 };
             }
         }
     };
-    ($fn:expr, $cost:expr, $execution_effect:expr, $transfers:expr) => {
+    ($fn:expr, $cost:expr, $execution_effect:expr, $execution_journal:expr, $transfers:expr) => {
         match $fn {
             Ok(res) => res,
             Err(e) => {
@@ -63,6 +64,7 @@ macro_rules! on_fail_charge {
                 return ExecutionResult::Failure {
                     error: exec_err.into(),
                     execution_effect: $execution_effect,
+                    execution_journal: $execution_journal,
                     transfers: $transfers,
                     cost: $cost,
                 };
@@ -208,6 +210,7 @@ impl Executor {
                 Ok(_value) => {
                     return ExecutionResult::Success {
                         execution_effect: runtime.context().effect(),
+                        execution_journal: runtime.context().execution_journal(),
                         transfers: runtime.context().transfers().to_owned(),
                         cost: runtime.context().gas_counter(),
                     };
@@ -216,6 +219,7 @@ impl Executor {
                     return ExecutionResult::Failure {
                         error: error.into(),
                         execution_effect,
+                        execution_journal: runtime.context().execution_journal(),
                         transfers: runtime.context().transfers().to_owned(),
                         cost: runtime.context().gas_counter(),
                     };
@@ -233,6 +237,7 @@ impl Executor {
                 Ok(_value) => {
                     return ExecutionResult::Success {
                         execution_effect: runtime.context().effect(),
+                        execution_journal: runtime.context().execution_journal(),
                         transfers: runtime.context().transfers().to_owned(),
                         cost: runtime.context().gas_counter(),
                     };
@@ -242,6 +247,7 @@ impl Executor {
                         error: error.into(),
                         execution_effect,
                         transfers: runtime.context().transfers().to_owned(),
+                        execution_journal: runtime.context().execution_journal(),
                         cost: runtime.context().gas_counter(),
                     };
                 }
@@ -257,6 +263,7 @@ impl Executor {
             ) {
                 Ok(_value) => {
                     return ExecutionResult::Success {
+                        execution_journal: runtime.context().execution_journal(),
                         execution_effect: runtime.context().effect(),
                         transfers: runtime.context().transfers().to_owned(),
                         cost: runtime.context().gas_counter(),
@@ -264,6 +271,7 @@ impl Executor {
                 }
                 Err(error) => {
                     return ExecutionResult::Failure {
+                        execution_journal: runtime.context().execution_journal(),
                         error: error.into(),
                         execution_effect,
                         transfers: runtime.context().transfers().to_owned(),
@@ -276,10 +284,12 @@ impl Executor {
             instance.invoke_export(entry_point_name, &[], &mut runtime),
             runtime.context().gas_counter(),
             execution_effect,
+            runtime.context().execution_journal(),
             runtime.context().transfers().to_owned()
         );
 
         ExecutionResult::Success {
+            execution_journal: runtime.context().execution_journal(),
             execution_effect: runtime.context().effect(),
             transfers: runtime.context().transfers().to_owned(),
             cost: runtime.context().gas_counter(),
@@ -349,6 +359,7 @@ impl Executor {
             Err(error) => {
                 return ExecutionResult::Failure {
                     error: error.into(),
+                    execution_journal: Default::default(),
                     execution_effect: Default::default(),
                     transfers: Vec::default(),
                     cost: Gas::default(),
@@ -360,11 +371,13 @@ impl Executor {
 
         match runtime.call_host_standard_payment() {
             Ok(()) => ExecutionResult::Success {
+                execution_journal: runtime.context().execution_journal(),
                 execution_effect: runtime.context().effect(),
                 transfers: runtime.context().transfers().to_owned(),
                 cost: runtime.context().gas_counter(),
             },
             Err(error) => ExecutionResult::Failure {
+                execution_journal: runtime.context().execution_journal(),
                 error: error.into(),
                 execution_effect,
                 transfers: runtime.context().transfers().to_owned(),
@@ -484,6 +497,7 @@ impl Executor {
         // Snapshot of effects before execution, so in case of error only nonce update
         // can be returned.
         let execution_effect = tracking_copy.borrow().effect();
+        let execution_journal = tracking_copy.borrow().execution_journal();
 
         let transfers = Vec::default();
 
@@ -512,6 +526,7 @@ impl Executor {
             Ok((instance, runtime)) => (instance, runtime),
             Err(error) => {
                 return ExecutionResult::Failure {
+                    execution_journal,
                     execution_effect,
                     transfers,
                     cost: gas_counter,
@@ -803,12 +818,14 @@ impl DirectSystemContractCall {
         match result {
             Ok(value) => match value.into_t() {
                 Ok(ret) => ExecutionResult::Success {
+                    execution_journal: runtime.context().execution_journal(),
                     execution_effect: runtime.context().effect(),
                     transfers: runtime.context().transfers().to_owned(),
                     cost: runtime.context().gas_counter(),
                 }
                 .take_with_ret(ret),
                 Err(error) => ExecutionResult::Failure {
+                    execution_journal: runtime.context().execution_journal(),
                     error: Error::CLValue(error).into(),
                     execution_effect,
                     transfers: runtime.context().transfers().to_owned(),
@@ -817,6 +834,7 @@ impl DirectSystemContractCall {
                 .take_without_ret(),
             },
             Err(error) => ExecutionResult::Failure {
+                execution_journal: runtime.context().execution_journal(),
                 error: error.into(),
                 execution_effect,
                 transfers: runtime.context().transfers().to_owned(),