+++ execution_engine/src/core/tracking_copy/mod.rs
@@ -29,6 +29,7 @@ use crate::{
     },
     shared::{
         additive_map::AdditiveMap,
+        execution_journal::ExecutionJournal,
         newtypes::CorrelationId,
         transform::{self, Transform},
     },
@@ -215,6 +216,7 @@ pub struct TrackingCopy<R> {
     cache: TrackingCopyCache<HeapSize>,
     ops: AdditiveMap<Key, Op>,
     fns: AdditiveMap<Key, Transform>,
+    journal: Vec<(Key, Transform)>,
 }
 
 #[derive(Debug)]
@@ -248,6 +250,7 @@ impl<R: StateReader<Key, StoredValue>> TrackingCopy<R> {
              * limit? */
             ops: AdditiveMap::new(),
             fns: AdditiveMap::new(),
+            journal: Vec::new(),
         }
     }
 
@@ -319,6 +322,7 @@ impl<R: StateReader<Key, StoredValue>> TrackingCopy<R> {
         if let Some(value) = self.get(correlation_id, &normalized_key)? {
             self.ops.insert_add(normalized_key, Op::Read);
             self.fns.insert_add(normalized_key, Transform::Identity);
+            self.journal.push((normalized_key, Transform::Identity));
             Ok(Some(value))
         } else {
             Ok(None)
@@ -329,7 +333,9 @@ impl<R: StateReader<Key, StoredValue>> TrackingCopy<R> {
         let normalized_key = key.normalize();
         self.cache.insert_write(normalized_key, value.clone());
         self.ops.insert_add(normalized_key, Op::Write);
-        self.fns.insert_add(normalized_key, Transform::Write(value));
+        let transform = Transform::Write(value);
+        self.fns.insert_add(normalized_key, transform.clone());
+        self.journal.push((normalized_key, transform));
     }
 
     /// Ok(None) represents missing key to which we want to "add" some value.
@@ -399,7 +405,8 @@ impl<R: StateReader<Key, StoredValue>> TrackingCopy<R> {
             Ok(new_value) => {
                 self.cache.insert_write(normalized_key, new_value);
                 self.ops.insert_add(normalized_key, Op::Add);
-                self.fns.insert_add(normalized_key, transform);
+                self.fns.insert_add(normalized_key, transform.clone());
+                self.journal.push((normalized_key, transform));
                 Ok(AddResult::Success)
             }
             Err(transform::Error::TypeMismatch(type_mismatch)) => {
@@ -411,6 +418,10 @@ impl<R: StateReader<Key, StoredValue>> TrackingCopy<R> {
 
     pub fn effect(&self) -> ExecutionEffect {
         ExecutionEffect::new(self.ops.clone(), self.fns.clone())
+    }
+
+    pub fn execution_journal(&self) -> ExecutionJournal {
+        self.journal.clone().into()
     }
 
     /// Calling `query()` avoids calling into `self.cache`, so this will not return any values