+++ execution_engine/src/shared.rs
@@ -1,6 +1,7 @@
 //! The shared logic of the execution engine.
 pub mod additive_map;
 #[macro_use]
+pub mod execution_journal;
 pub mod host_function_costs;
 pub mod logging;
 pub mod newtypes;